feat: add strategies in an stand alone json file
