+++ src/views/Position/Detail.tsx
@@ -2,7 +2,7 @@ import BoxWrapperColumn from 'src/components/Wrappers/BoxWrapperColumn'
 import * as React from 'react'
 import { PositionType } from 'src/contexts/types'
 import { Divider } from '@mui/material'
-import { getStrategies, DAO } from 'src/config/strategies'
+import { getStrategies, DAO } from 'src/config/strategiesManager'
 import Form from 'src/views/Position/Form'
 import Primary from 'src/views/Position/Title/Primary'
 import Secondary from 'src/views/Position/Title/Secondary'
